Added config option and support for max children
Summary:
To prevent fork-bombing a box due to poor client
behavior, added in a max children default of 512 and a
config option to set it.

There is also a maxprocsperclt setting (default 16) that
will control the max number of processes a specific client
ip can open.

The socket timeouts have also been lowered from 180 secs to
15 secs.