Add authorization logging, fix double accept condition, use sockaddr_storage instead of sockaddr_in
Adds authorization logging and flag to turn on/off
Fixes condition under high load where a v4 and v6 request may be accepted and forked to the same child... results in one fd being stranded and the other being serviced as normal. Breaking the for loop is not really the ideal way to fix it, however it beats refactoring all of main() at the moment.

Uses sockaddr_storage instead of sockaddr_in to be address family agnostic